added dot reporter
